+++ src/api/util/index.ts
@@ -1,2 +1,3 @@
 export { isNonNegativeNumber } from './isNonNegativeNumber';
 export { isPercent } from './isPercent';
+export { isPositiveLength } from './isPositiveLength';