+++ src/api/util/isNonNegativeNumber.ts
@@ -1,6 +1,8 @@
 /**
  * The `isNonNegativeNumber` method returns whether the given value is a non-negative number.
  *
+ * ([0-9]+(\.[0-9]*)?|\.[0-9]+)((cm)|(mm)|(in)|(pt)|(pc)|(px))
+ *
  * @param {any} value The value that is to be checked
  * @returns {boolean} `true` if the given value is a non-negative number, `false` otherwise
  * @private